Vineri V1
Versiune luata de la Paul de pe laptop.

- UART ok

+++ SAM3X8E/GccApplication1/MotorCtrl.h
@@ -12,6 +12,6 @@
 	void InitMotors(void);
 	void WriteMotors(int percent_ST, int percent_DR);
 	void ForwardDrive(void);
-	void ControlledDrive(uint8_t percentage_ST, uint8_t percentage_DR);
+	void ControlledDrive(uint32_t percentage_ST, uint32_t percentage_DR);
 	
 #endif
